requirements/main.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/main.txt requirements/main.in
#
aiohttp==3.7.4.post0
    # via
    #   -r requirements/main.in
    #   graphql-server
    #   safir
astropy==4.2
    # via
    #   -r requirements/main.in
    #   daf-butler
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
chardet==4.0.0
    # via aiohttp
click==7.1.2
    # via
    #   -r requirements/main.in
    #   daf-butler
git+git://github.com/lsst/daf_butler.git@master#daf_butler
    # via -r requirements/main.in
deprecated==1.2.12
    # via daf-butler
graphql-core==3.1.3
    # via graphql-server
graphql-server[aiohttp]==3.0.0b3
    # via -r requirements/main.in
greenlet==1.0.0
    # via sqlalchemy
idna==3.1
    # via yarl
importlib-metadata==2.1.1
    # via -r requirements/main.in
jinja2==2.11.3
    # via -r requirements/main.in
git+https://github.com/lsst/sphgeom@master
    # via daf-butler
markupsafe==1.1.1
    # via jinja2
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.20.2
    # via
    #   astropy
    #   lsst-sphgeom
    #   pyerfa
pyerfa==1.7.2
    # via astropy
pyyaml==5.4.1
    # via daf-butler
safir==0.1.1
    # via -r requirements/main.in
sqlalchemy==1.4.3
    # via daf-butler
structlog==21.1.0
    # via safir
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   graphql-server
wrapt==1.12.1
    # via deprecated
yarl==1.6.3
    # via aiohttp
zipp==3.4.1
    # via importlib-metadata
